Refactor : Applied Go's basic project structure
